roles collection of rest api
